+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.1.0] - 2025-11-20
+
+## Added
+- Add action to update binary (binary auto update command)
+
+## Updated
+- ngrok binary (v3.33.0)
+
 ## [1.0.1] - 2024-11-22
 
 ## Updated